Move gc_layout intrinsic logic into src/ty
This lets rustc call `gc_layout` internally as well as it being an
intrinsic through libcore.
